requirements.txt
gradio_imageslider
gradio
matplotlib
torch
torchvision
streamlit
streamlit-drawable-canvas
opencv-python-headless
numpy
Pillow
fpdf
plotly
rasterio
scikit-image
fpdf2
pyvista
stpyvista
streamlit-echarts
requests
streamlit-autorefresh
streamlit-image-coordinates
streamlit-webrtc
streamlit-lottie
